+++ changelogs/fragments/54124-docker_image-facts-info.yml
@@ -1,0 +1,2 @@
+minor_changes:
+- The ``docker_image_facts`` module has been renamed to ``docker_image_info``.
