+++ lib/ansible/modules/cloud/docker/_docker_image_facts.py
@@ -1,0 +1,1 @@
+docker_image_info.py
