+++ test/integration/targets/docker_container_info/aliases
