+++ test/integration/targets/docker_container_info/meta/main.yml
