+++ test/integration/targets/docker_host_info/aliases
