+++ test/integration/targets/docker_host_info/meta/main.yml
