+++ test/integration/targets/docker_host_info/tasks/main.yml
@@ -1,0 +1,5 @@
+- include_tasks: test_host_info.yml
+  when: docker_py_version is version('1.10.0', '>=') and docker_api_version is version('1.21', '>=')
+
+- fail: msg="Too old docker / docker-py version to run docker_host_info tests!"
+  when: not(docker_py_version is version('1.10.0', '>=') and docker_api_version is version('1.21', '>=')) and (ansible_distribution != 'CentOS' or ansible_distribution_major_version|int > 6)
