+++ test/integration/targets/docker_image_info/aliases
