+++ test/integration/targets/docker_image_info/meta/main.yml
