+++ test/integration/targets/docker_network_info/aliases
