+++ test/integration/targets/docker_node_info/aliases
