+++ test/integration/targets/docker_node_info/meta/main.yml
