+++ test/integration/targets/docker_swarm_info/aliases
