+++ test/integration/targets/docker_swarm_info/meta/main.yml
