+++ test/integration/targets/docker_volume_info/aliases
