+++ test/integration/targets/docker_volume_info/meta/main.yml
